refactor(studio): extract IntakeTraceDetailBody to components dir (FP-185)
- Move IntakeTraceDetailBody and its props interface out of
  IntakeTraceDetailRoute into @studio/components/IntakeTraceDetailBody
- Remove redundant useGetTrace call from IntakeTraceDetailContent;
  use traceId directly for breadcrumb/heading until trace data loads
- Update ExperimentSessionsDataView import to new component location
